+++ app/models/unidom/party/concerns/as_source_party.rb
@@ -1,0 +1,14 @@
+module Unidom::Party::Concerns::AsSourceParty
+
+  extend ActiveSupport::Concern
+
+  included do |includer|
+
+    has_many :target_party_relations, class_name: 'Unidom::Party::PartyRelation', as: :source_party
+
+  end
+
+  module ClassMethods
+  end
+
+end